android_jni: Unify tests in a single file
Eventhough the tests were separated as still and animated, the
actual separation was about whether they were using the Basic
API or the Full API.

Unify the tests and subject all the files to both the Basic API
and the Full API.
