+++ android_jni/avifandroidjni/src/androidTest/proguard-rules.pro
@@ -2,9 +2,6 @@
 
 # These classes are needed by the instrumentation testing framework. Keep them
 # unobfuscated.
--keep class org.aomedia.avif.android.StillImageTest {
-  *;
-}
--keep class org.aomedia.avif.android.AnimatedImageTest {
+-keep class org.aomedia.avif.android.AvifDecoderTest {
   *;
 }
